requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
cachetools==6.1.0
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.4
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-cities-light==3.10.1
django-ip-geolocation==1.6.1
django-ipware==7.0.1
django-ratelimit==4.1.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
dotenv==0.9.9
frozenlist==1.7.0
hiredis==3.2.1
idna==3.10
ipinfo==4.2.1
ipinfo-django==2.0.0
joblib==1.5.1
kombu==5.5.4
multidict==6.6.3
numpy==2.2.6
packaging==25.0
pandas==2.3.1
progressbar2==4.5.0
prompt_toolkit==3.0.51
propcache==0.3.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
python-utils==3.9.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
